refactor: programmer 디렉토리명을 service로 변경
기존에는 `ProgrammerService` 비즈니스 로직만을 검증하기 위한 클래스였기 때문에 디렉토리명이 `programmer`였지만 그렇게 하면 테스트 코드가 클래스마다 너무 작게 들어가기 때문에 하나의 디렉토리로 합치게 되었고 디렉토리명도 `service`로 변경하게 되었다.
